Altered a few things.
Hello, I'm new to Github, I've taken this code and have gotten to work with Python 3.9. Admittedly I'm a novice to this and am looking for problems to fix to get experience. This just happened to be the first project I came across. If I haven't done this correctly or if you have any tips for a beginner please let me know. Just trying to do what I can to help around here.